requirements.txt
# requirements.txt
pip>=23.0
dbt-snowflake>=1.5.0
prefect>=2.0.0
snowflake-connector-python>=3.0.0
scikit-learn>=1.3.0
mlflow>=2.5.0
shap>=0.42.0
pandas>=2.0.0
pyarrow>=12.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyterlab>=4.0.0